requirements.txt
cachetools==5.2.0
google-auth==2.14.0
google-auth-oauthlib==0.7.1
gspread==5.6.2
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.1
rsa==4.9
